requirements.dev.txt
codecov
pytest-cov
pylint
mock
jsonschema
numpy
